v3 Actors Amt update (#978)
* Switch cache to OnceCell

* Remove lazycell from hamt

* Replace state manager lazycell

* Cleanup init in amt

* cleanup vesting initialization

* Removing old version caching inefficiencies

* Update batch delete

* wip new functionality

* Cleanup and bug fixes

* fiz fuzzer

* reloading flushed amt for more converage in fuzzer

* Cleanup and replace with released version

* Add changelog and bump version for release

* bump actors version and fix comments

+++ blockchain/chain/src/store/chain_store.rs
@@ -922,13 +922,13 @@ where
 pub fn get_parent_reciept<DB>(
     db: &DB,
     block_header: &BlockHeader,
-    i: u64,
+    i: usize,
 ) -> Result<Option<MessageReceipt>, Error>
 where
     DB: BlockStore,
 {
     let amt = Amt::load(block_header.message_receipts(), db)?;
-    let receipts = amt.get(i)?;
+    let receipts = amt.get(i as u64)?;
     Ok(receipts.cloned())
 }
